This version fixes a bug in the acquisition of patient IDs having a given itemset of comorbidity conditions.
In the previous version, there was an accurate count of the number of
patients having a given comorbidity condition.  Supposing that 100
patients had all of the conditions in the itemset, the first 99 were
accurately identified, but the patient ID of the final one was incorrectly
identified.  The edited file now fixes this.

TS